requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
atproto==0.0.56
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
dnspython==2.7.0
exceptiongroup==1.2.2
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huepy==1.2.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
instabot==0.117.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
libipld==3.0.0
mock==5.1.0
moviepy==2.1.1
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.4
orjson==3.10.12
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.2.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.0
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
python-linkedin-v2==0.9.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
